+++ src/cpp/session/modules/SessionSource.cpp
@@ -1,7 +1,7 @@
 /*
  * SessionSource.cpp
  *
- * Copyright (C) 2009-12 by RStudio, Inc.
+ * Copyright (C) 2009-18 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -146,7 +146,7 @@ int numSourceDocuments()
 {
    std::vector<boost::shared_ptr<SourceDocument> > docs;
    source_database::list(&docs);
-   return docs.size();
+   return static_cast<int>(docs.size());
 }
 
 // wrap source_database::put for situations where there are new contents
